refactor: Move entities extractors to dedicated class
